Change disposal to allow for pooling of pools according to stream
