+++ src/Speckle.Sdk/SQLite/SQLiteJsonCacheManager.cs
@@ -5,93 +5,13 @@ using Speckle.Sdk.Common;
 using Speckle.Sdk.Dependencies;
 
 namespace Speckle.Sdk.SQLite;
-
 public partial interface ISqLiteJsonCacheManager : IDisposable;
-
 [GenerateAutoInterface]
-public sealed class SqLiteJsonCacheManager : ISqLiteJsonCacheManager
+public sealed class SqLiteJsonCacheManager(ISqliteJsonCachePool pool, bool dispose) : ISqLiteJsonCacheManager
 {
-  private readonly CacheDbCommandPool _pool;
-
-  public SqLiteJsonCacheManager(string path, int concurrency)
-  {
-    Path = path;
-    Concurrency = concurrency;
-    //disable pooling as we pool ourselves
-    var builder = new SqliteConnectionStringBuilder { Pooling = false, DataSource = path };
-    _pool = new CacheDbCommandPool(builder.ToString(), concurrency);
-    Initialize();
-  }
   
-  public string Path { get; }
-  public int Concurrency { get; }
-
-  [AutoInterfaceIgnore]
-  public void Dispose() => _pool.Dispose();
-
-  private void Initialize()
-  {
-    // NOTE: used for creating partioned object tables.
-    //string[] HexChars = new string[] { "0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "a", "b", "c", "d", "e", "f" };
-    //var cart = new List<string>();
-    //foreach (var str in HexChars)
-    //  foreach (var str2 in HexChars)
-    //    cart.Add(str + str2);
-
-    _pool.Use(c =>
-    {
-      const string COMMAND_TEXT =
-        @"
-            CREATE TABLE IF NOT EXISTS objects(
-              hash TEXT PRIMARY KEY,
-              content TEXT
-            ) WITHOUT ROWID;
-          ";
-      using (var command = new SqliteCommand(COMMAND_TEXT, c))
-      {
-        command.ExecuteNonQuery();
-      }
-
-      // Insert Optimisations
-
-      //Note / Hack: This setting has the potential to corrupt the db.
-      //cmd = new SqliteCommand("PRAGMA synchronous=OFF;", Connection);
-      //cmd.ExecuteNonQuery();
-
-      using (SqliteCommand cmd1 = new("PRAGMA count_changes=OFF;", c))
-      {
-        cmd1.ExecuteNonQuery();
-      }
-
-      using (SqliteCommand cmd2 = new("PRAGMA temp_store=MEMORY;", c))
-      {
-        cmd2.ExecuteNonQuery();
-      }
-
-      using (SqliteCommand cmd3 = new("PRAGMA mmap_size = 30000000000;", c))
-      {
-        cmd3.ExecuteNonQuery();
-      }
-
-      using (SqliteCommand cmd4 = new("PRAGMA page_size = 32768;", c))
-      {
-        cmd4.ExecuteNonQuery();
-      }
-
-      using (SqliteCommand cmd5 = new("PRAGMA journal_mode='wal';", c))
-      {
-        cmd5.ExecuteNonQuery();
-      }
-      //do this to wait 5 seconds to avoid db lock exceptions, this is 0 by default
-      using (SqliteCommand cmd6 = new("PRAGMA busy_timeout=5000;", c))
-      {
-        cmd6.ExecuteNonQuery();
-      }
-    });
-  }
-
   public IReadOnlyCollection<(string Id, string Json)> GetAllObjects() =>
-    _pool.Use(
+    pool.Use(
       CacheOperation.GetAll,
       command =>
       {
@@ -106,7 +26,7 @@ public sealed class SqLiteJsonCacheManager : ISqLiteJsonCacheManager
     );
 
   public void DeleteObject(string id) =>
-    _pool.Use(
+    pool.Use(
       CacheOperation.Delete,
       command =>
       {
@@ -116,7 +36,7 @@ public sealed class SqLiteJsonCacheManager : ISqLiteJsonCacheManager
     );
 
   public string? GetObject(string id) =>
-    _pool.Use(
+    pool.Use(
       CacheOperation.Get,
       command =>
       {
@@ -130,7 +50,7 @@ public sealed class SqLiteJsonCacheManager : ISqLiteJsonCacheManager
   {
     id.NotNullOrWhiteSpace();
     json.NotNullOrWhiteSpace();
-    _pool.Use(
+    pool.Use(
       CacheOperation.InsertOrIgnore,
       command =>
       {
@@ -143,7 +63,7 @@ public sealed class SqLiteJsonCacheManager : ISqLiteJsonCacheManager
 
   //This does an insert or replaces if already exists
   public void UpdateObject(string id, string json) =>
-    _pool.Use(
+    pool.Use(
       CacheOperation.InsertOrReplace,
       command =>
       {
@@ -154,7 +74,7 @@ public sealed class SqLiteJsonCacheManager : ISqLiteJsonCacheManager
     );
 
   public void SaveObjects(IEnumerable<(string id, string json)> items) =>
-    _pool.Use(
+    pool.Use(
       CacheOperation.BulkInsertOrIgnore,
       cmd =>
       {
@@ -200,7 +120,7 @@ public sealed class SqLiteJsonCacheManager : ISqLiteJsonCacheManager
   }
 
   public bool HasObject(string objectId) =>
-    _pool.Use(
+    pool.Use(
       CacheOperation.Has,
       command =>
       {
@@ -210,4 +130,13 @@ public sealed class SqLiteJsonCacheManager : ISqLiteJsonCacheManager
         return rowFound;
       }
     );
+
+  [AutoInterfaceIgnore]
+  public void Dispose()
+  {
+    if (dispose)
+    {
+      pool.Dispose();
+    }
+  }
 }
